+++ packages/openapi/src/ui/components/input.tsx
@@ -5,7 +5,7 @@ import { cva } from 'class-variance-authority';
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
 
 export const labelVariants = cva(
-  'text-[13px] font-mono text-fd-card-foreground peer-disabled:cursor-not-allowed peer-disabled:opacity-70',
+  'text-[13px] font-medium text-fd-card-foreground peer-disabled:cursor-not-allowed peer-disabled:opacity-70',
 );
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(